+++ tests/models/test_models_shef.py
@@ -43,7 +43,7 @@ def test_lonlat(elem):
     assert abs(res[0] - -99.5) < 0.01
     assert abs(res[1] - -30.8) < 0.01
 
-    elem.station = "W30800994A"
+    elem.station = "W308099"
     res = elem.lonlat()
     assert res[0] is None