+++ CHANGELOG.md
@@ -8,6 +8,7 @@ All notable changes to this library are documented in this file.
 ### API Changes
 
 - Enforce a SHEF data model limit of 8 characters for a station identifier.
+- Ween off `domain=` nomenclature for the default IEMRE domain.
 
 ### New Features